probe/fbconfig: Texture-map a quadrangle using the "legacy" APIs
  * This was... obnoxious to figure out.  Need to set up a small
pile of state in the context, with very little in the way of
diagnostic advice when all you get for output is a white rectangle
and no texture.

  * The legacy (no shader!) APIs are based on the fixed-function
pipeline, and apparently not supported by certain OpenGL
implementations such as the Intel open-source Linux drivers.

  * Apparently, the "modern" way to do things is to set up a
"vertex buffer", store the coordinates there, and then use shaders
to do the geometry mangling and texture mapping.